+++ code/src/cljs/sixsq/nuvla/ui/credentials/views.cljs
@@ -13,7 +13,6 @@
     [sixsq.nuvla.ui.main.components :as main-components]
     [sixsq.nuvla.ui.panel :as panel]
     [sixsq.nuvla.ui.session.subs :as session-subs]
-    [sixsq.nuvla.ui.utils.accordion :as utils-accordion]
     [sixsq.nuvla.ui.utils.general :as general-utils]
     [sixsq.nuvla.ui.utils.semantic-ui :as ui]
     [sixsq.nuvla.ui.utils.semantic-ui-extensions :as uix]
@@ -353,48 +352,25 @@
       {:on-refresh #(dispatch [::events/get-credentials])}]]))
 
 
-(defn delete-confirmation-modal
-  []
-  (let [tr         (subscribe [::i18n-subs/tr])
-        visible?   (subscribe [::subs/delete-confirmation-modal-visible?])
-        credential (subscribe [::subs/credential])
-        confirmed? (r/atom false)]
-    (fn []
-      (let [id   (:id @credential)
-            name (:name @credential)]
-        [ui/Modal {:open       @visible?
-                   :close-icon true
-                   :on-close   #(dispatch [::events/close-delete-confirmation-modal])}
-
-         [ui/ModalHeader (str "Delete Credential?")]
-
-         [ui/ModalContent {:scrolling false}
-          [:h3 name " (" (:description @credential) ")"]
-          [:div "(" id ")"]
-          [ui/Message {:error true}
-           [ui/MessageHeader "Danger - this cannot be undone!"]
-           [ui/MessageContent
-            [:p]
-            [ui/Checkbox {:name      "confirm-deletion"
-                          :label     (@tr [:credential-delete-warning])
-                          :checked   @confirmed?
-                          :fitted    true
-                          :on-change #(reset! confirmed? (not @confirmed?))}]]]]
-
-         [ui/ModalActions
-          [uix/Button {:text     (@tr [:delete])
-                       :positive true
-                       :disabled (when-not @confirmed? true)
-                       :active   true
-                       :on-click #(do
-                                    (reset! confirmed? false)
-                                    (dispatch [::events/delete-credential id])
-                                    (dispatch [::events/close-delete-confirmation-modal]))}]]]))))
+(defn DeleteButton
+  [credential]
+  (let [tr (subscribe [::i18n-subs/tr])
+        {:keys [id name description]} credential
+        content (str (or name id) (when description " - ") description)]
+    [uix/ModalDanger
+     {:on-confirm  #(dispatch [::events/delete-credential id])
+      :trigger     (r/as-element [ui/Icon {:name  "trash"
+                                           :style {:cursor "pointer"}
+                                           :color "red"}])
+      :content     [:h3 content]
+      :header      (@tr [:delete-credential])
+      :danger-msg  (@tr [:credential-delete-warning])
+      :button-text (@tr [:delete])}]))
 
 
 ;subtype name description
 (defn single-credential
-  [{:keys [id subtype name description] :as credential}]
+  [{:keys [subtype name description] :as credential}]
   [ui/TableRow
    [ui/TableCell {:floated :left
                   :width   2}
@@ -411,7 +387,7 @@
                   :style   {}}
 
     (when (general-utils/can-delete? credential)
-      [utils-accordion/trash id ::events/open-delete-confirmation-modal nil credential])
+      [DeleteButton credential])
 
     (when (general-utils/can-edit? credential)
       [ui/Icon {:name     :cog
@@ -424,10 +400,10 @@
   []
   (let [tr          (subscribe [::i18n-subs/tr])
         credentials (subscribe [::subs/credentials])]
+    (dispatch [::events/get-credentials])
     (fn []
       (let [infra-service-creds (filter #(in? infrastructure-service-subtypes (:subtype %))
                                         @credentials)]
-        (dispatch [::events/get-credentials])
         [ui/Container {:fluid true}
          [uix/PageHeader "key" (str/capitalize (@tr [:credentials])) :inline true]
          [uix/Accordion
@@ -459,5 +435,4 @@
    [credentials]
    [add-credential-modal]
    [credential-modal]
-   [generated-credential-modal]
-   [delete-confirmation-modal]])
+   [generated-credential-modal]])